Refactor implementation to separate file
